Lana/collision

+++ game/src/gui/console/Console.cpp
@@ -1,21 +1,26 @@
 
+#include <fornani/events/ConsoleEvent.hpp>
+#include <fornani/events/GameplayEvent.hpp>
+#include <fornani/events/InventoryEvent.hpp>
+#include <fornani/events/SystemEvent.hpp>
 #include <fornani/gui/console/Console.hpp>
-
-#include <fornani/systems/Event.hpp>
 #include "fornani/service/ServiceProvider.hpp"
-#include "fornani/setup/ControllerMap.hpp"
+#include "fornani/systems/InputSystem.hpp"
 
 namespace fornani::gui {
 
 Console::Console(automa::ServiceProvider& svc)
 	: m_services(&svc), m_path{svc.finder, std::filesystem::path{"/data/gui/console_paths.json"}, "standard", 64}, m_styling{.corner_factor{28}, .edge_factor{1}, .padding_scale{1.1f}},
-	  m_nineslice(svc, svc.assets.get_texture("blue_console"), {m_styling.corner_factor, m_styling.corner_factor}, {m_styling.edge_factor, m_styling.edge_factor}), m_mode{ConsoleMode::writing}, m_response_offset{-128.f, 16.f},
-	  m_exit_stall{650} {
+	  m_nineslice(svc, svc.assets.get_texture("blue_console"), {m_styling.corner_factor, m_styling.corner_factor}, {m_styling.edge_factor, m_styling.edge_factor}), m_mode{ConsoleMode::writing}, m_response_offset{-192.f, 16.f},
+	  m_exit_stall{450}, m_item_display_timer{1200}, m_launch{8} {
+	NANI_LOG_DEBUG(m_logger, "Console ctor @{}", static_cast<void const*>(this));
+	svc.input_system.set_action_set(input::ActionSet::Menu);
 	text_suite = svc.text.console;
 	m_path.set_section("open");
 	m_began = true;
 	m_path.update();
 	m_position = m_path.get_position();
+	m_launch.start();
 }
 
 Console::Console(automa::ServiceProvider& svc, std::string_view message) : Console(svc) { load_single_message(message); }
@@ -26,24 +31,35 @@ Console::Console(automa::ServiceProvider& svc, dj::Json const& source, OutputTyp
 	load_and_launch(type);
 }
 
-Console::Console(automa::ServiceProvider& svc, dj::Json const& source, std::string_view key, OutputType type) : Console(svc) {
+Console::Console(automa::ServiceProvider& svc, dj::Json const& source, std::string_view key, OutputType type, int target_index) : Console(svc) {
 	if (type == OutputType::no_skip) { m_exit_stall.start(); }
 	set_source(source);
-	load_and_launch(key, type);
+	load_and_launch(key, type, target_index);
+}
+
+void Console::relaunch(automa::ServiceProvider& svc, dj::Json const& source, std::string_view key, OutputType type, int target_index) {
+	m_mode = ConsoleMode::writing;
+	NANI_LOG_DEBUG(m_logger, "Relaunched console.");
+	if (type == OutputType::no_skip) { m_exit_stall.start(); }
+	set_source(source);
+	load_and_launch(key, type, target_index);
 }
 
 void Console::update(automa::ServiceProvider& svc) {
 	m_exit_stall.update();
+	m_launch.update();
+	m_item_display_timer.update();
+	if (m_item_display_timer.is_almost_complete()) { m_item_widget->send_out(); }
 	if (!is_active()) { return; }
 	if (!m_writer) { return; }
 	m_began = false;
-	handle_inputs(svc.controller_map);
+	handle_inputs(svc.input_system);
 	if (is_active()) { m_path.update(); }
 	if (m_path.finished() && m_writer) {
 		if (m_writer->is_stalling()) { m_writer->start(); }
 	}
 	if (!m_writer) { return; }
-	if (m_writer->is_writing()) { svc.soundboard.flags.console.set(audio::Console::speech); }
+	if (m_writer->is_writing()) { svc.soundboard.repeat_sound("console_speech", 0, {}, m_writer->is_sped_up() ? 1.6f : 1.f); }
 	m_position = m_path.get_position();
 	m_dimensions = m_path.get_dimensions();
 	m_nineslice.set_offset(-m_path.get_local_center());
@@ -64,10 +80,24 @@ void Console::update(automa::ServiceProvider& svc) {
 	if (auto codes = get_message_codes()) {
 		for (auto& code : codes.value()) {
 			if (code.is_response() && m_writer->is_available()) { m_writer->respond(); }
+			if (code.is_reveal_item() && m_process_code_before) {
+				m_services->events.reveal_item_by_id_event.dispatch(code.value);
+				processed = true;
+			}
 			if (code.is_input_hint()) {
 				auto action_id = code.extras ? code.extras->at(0) : 0;
-				auto lookup = m_services->controller_map.get_icon_lookup_by_action(static_cast<config::DigitalAction>(action_id));
+				auto lookup = m_services->input_system.get_icon_lookup_by_action(static_cast<input::DigitalAction>(action_id));
 				m_writer->insert_icon_at(code.value, lookup);
+			}
+			if (code.is(MessageCodeType::launch_cutscene) && m_process_code_before) {
+				m_services->events.launch_cutscene_event.dispatch(*m_services, code.value);
+				processed = true;
+			}
+			if (code.is(MessageCodeType::add_map_marker) && m_process_code_before && code.extras) {
+				if (code.extras->size() > 1) {
+					m_services->events.add_map_marker_event.dispatch(*m_services, code.value, code.extras->at(0), code.extras->at(1));
+					processed = true;
+				}
 			}
 			if (code.is_quest() && m_process_code_before) {
 				if (code.extras) {
@@ -76,7 +106,7 @@ void Console::update(automa::ServiceProvider& svc) {
 				processed = true;
 			}
 			if (code.is_pop_conversation() && m_process_code_before) {
-				svc.events.dispatch_event("PopConversation", code.value);
+				m_services->events.npc_pop_conversation_event.dispatch();
 				auto label = svc.data.get_npc_label_from_id(code.value);
 				if (label && code.extras) {
 					svc.quest_table.progress_quest("npc_dialogue", {label.value().data(), code.extras->at(0)}, 1, -1, code.value);
@@ -84,32 +114,26 @@ void Console::update(automa::ServiceProvider& svc) {
 				}
 				processed = true;
 			}
-			if (code.is_open_vendor()) { m_services->events.dispatch_event("OpenVendor", code.value); }
 			if (code.is_play_song() && m_process_code_before) {
-				svc.events.dispatch_event("PlaySong", code.value);
+				m_services->events.play_song_event.dispatch(code.value);
 				processed = true;
 			}
 			if (code.is_voice_cue() && m_process_code_before) {
-				svc.events.dispatch_event("VoiceCue", code.value);
-				NANI_LOG_DEBUG(m_logger, "Voice!");
+				// m_services->events.get_or_add<NPCVoiceCueEvent>().dispatch(*m_services, code.value);
+				m_services->events.npc_voice_cue_event.dispatch(*m_services, code.value);
+				NANI_LOG_DEBUG(m_logger, "Voice Cue: {}", code.value);
 				processed = true;
 			}
 			if (code.is_emotion() && m_process_code_before && m_npc_portrait) {
 				m_npc_portrait->set_emotion(code.value);
 				if (code.extras) {
 					if (!code.extras->empty()) {
-						if (code.extras->at(0) == 1) {
-							m_npc_portrait->add_sparkler("portrait_sparkle");
-							NANI_LOG_DEBUG(m_logger, "Added a sparkler.");
-						}
-					} else {
-						NANI_LOG_DEBUG(m_logger, "Uh oh... code.extras was empty!");
+						if (code.extras->at(0) == 1) { m_npc_portrait->add_sparkler("portrait_sparkle"); }
 					}
 				} else {
 					m_npc_portrait->remove_sparkler();
-					NANI_LOG_DEBUG(m_logger, "Removed the sparkler.");
 				}
-				NANI_LOG_DEBUG(m_logger, "Emotion!");
+				NANI_LOG_DEBUG(m_logger, "Emotion: {}", code.value);
 				processed = true;
 			}
 		}
@@ -121,6 +145,7 @@ void Console::update(automa::ServiceProvider& svc) {
 }
 
 void Console::render(sf::RenderWindow& win) {
+	if (m_launch.running()) { return; }
 	// debug();
 	if (!m_writer || !is_active()) { return; }
 	m_nineslice.render(win);
@@ -136,19 +161,32 @@ void Console::render(sf::RenderWindow& win) {
 
 void Console::set_source(dj::Json const& json) { text_suite = json; }
 
-void Console::set_nani_sprite(sf::Sprite const& sprite) { m_nani_portrait = Portrait(*m_services, sprite.getTexture(), 5, false); }
+void Console::set_nani_sprite(sf::Sprite const& sprite) {
+	if (m_response) {
+		if (m_response->has_flag_set(ResponseDialogFlags::hide_portrait)) { return; }
+	}
+	m_nani_portrait = Portrait(*m_services, sprite.getTexture(), 5, false);
+}
 
 void Console::handle_actions(int value) {
 	switch (value) {
 	case 1: // return to main menu
-		m_services->state_controller.actions.set(automa::Actions::main_menu);
-		m_services->state_controller.actions.set(automa::Actions::trigger);
+		NANI_LOG_DEBUG(m_logger, "Firing event from console: ReturnToMainMenuEvent");
+		m_services->events.return_to_main_menu_event.dispatch();
 		break;
 	case 2: // restart save
-		m_services->state_controller.actions.set(automa::Actions::retry);
+		NANI_LOG_DEBUG(m_logger, "Firing event from console: ReloadSaveEvent");
+		m_services->events.reload_save_event.dispatch(*m_services, value);
 		break;
 	case 3: m_services->state_controller.actions.set(automa::Actions::delete_file); break;
 	}
+}
+
+void Console::load_and_launch(std::string_view key, OutputType type, int target_index) {
+	m_writer = std::make_unique<TextWriter>(*m_services, text_suite, key, target_index);
+	m_process_code_before = true;
+	m_output_type = type;
+	native_key = key;
 }
 
 void Console::load_and_launch(OutputType type) {
@@ -156,21 +194,20 @@ void Console::load_and_launch(OutputType type) {
 	m_process_code_before = true;
 	m_output_type = type;
 	native_key = null_key;
-	update(*m_services);
-}
-
-void Console::load_and_launch(std::string_view key, OutputType type) {
-	m_writer = std::make_unique<TextWriter>(*m_services, text_suite, key);
-	m_process_code_before = true;
-	m_output_type = type;
-	native_key = key;
 }
 
 void Console::load_single_message(std::string_view message) { m_writer = std::make_unique<TextWriter>(*m_services, message); }
 
-void Console::display_item(int item_id) { m_item_widget = ItemWidget(*m_services, ItemWidgetType::item, item_id); }
+void Console::display_item(std::string_view tag, bool sparkle) {
+	m_item_widget = ItemWidget(*m_services, ItemWidgetType::item, tag);
+	m_item_display_timer.start();
+	if (!sparkle) { m_item_widget->remove_sparkler(); }
+}
 
-void Console::display_gun(int gun_id) { m_item_widget = ItemWidget(*m_services, ItemWidgetType::gun, gun_id); }
+void Console::display_gun(std::string_view tag, bool sparkle) {
+	m_item_widget = ItemWidget(*m_services, ItemWidgetType::gun, tag);
+	if (!sparkle) { m_item_widget->remove_sparkler(); }
+}
 
 void Console::write(sf::RenderWindow& win, bool instant) {
 	if (!is_active()) { return; }
@@ -183,6 +220,7 @@ void Console::write(sf::RenderWindow& win) { write(win, m_output_type == OutputT
 void Console::append(std::string_view key) { m_writer->append(key); }
 
 void Console::end() {
+	NANI_LOG_DEBUG(m_logger, "Ended console.");
 	m_services->soundboard.flags.console.set(audio::Console::done);
 	m_mode = ConsoleMode::off; // handle exit codes
 	m_writer = {};
@@ -192,16 +230,16 @@ void Console::include_portrait(int id) { m_npc_portrait = Portrait{*m_services, 
 
 std::string Console::get_key() const { return native_key; }
 
-void Console::handle_inputs(config::ControllerMap& controller) {
+void Console::handle_inputs(input::InputSystem& controller) {
 	m_triggers = {};
 	m_process_code_after = false;
 	if (m_exit_stall.running()) { return; }
-	auto const& up = controller.digital_action_status(config::DigitalAction::menu_up).triggered;
-	auto const& down = controller.digital_action_status(config::DigitalAction::menu_down).triggered;
-	auto const& next = controller.digital_action_status(config::DigitalAction::menu_select).triggered;
-	auto const& exit = controller.digital_action_status(config::DigitalAction::menu_cancel).triggered && !m_flags.test(ConsoleFlags::no_exit);
-	auto const& skip = controller.digital_action_status(config::DigitalAction::menu_select).held;
-	auto const& released = controller.digital_action_status(config::DigitalAction::menu_select).released;
+	auto const& up = controller.menu_move(input::MoveDirection::up);
+	auto const& down = controller.menu_move(input::MoveDirection::down);
+	auto const& next = controller.digital(input::DigitalAction::menu_select).triggered;
+	auto const& exit = controller.digital(input::DigitalAction::menu_back).triggered && !m_flags.test(ConsoleFlags::no_exit);
+	auto const& skip = controller.digital(input::DigitalAction::menu_select).held;
+	auto const& released = controller.digital(input::DigitalAction::menu_select).released;
 	static bool finished{};
 	static bool can_skip{};
 	static bool just_started{};
@@ -229,8 +267,9 @@ void Console::handle_inputs(config::ControllerMap& controller) {
 			if (auto response_codes = get_response_codes(m_response->get_selection())) {
 				for (auto const& cde : response_codes.value()) {
 					if (cde.is_response()) { m_writer->set_suite(cde.value); }
+					if (cde.is_start_battle()) { m_services->events.start_battle_event.dispatch(); }
 					if (cde.is_pop_conversation()) {
-						m_services->events.dispatch_event("PopConversation", cde.value);
+						m_services->events.npc_pop_conversation_event.dispatch();
 						auto label = m_services->data.get_npc_label_from_id(cde.value);
 						if (label && cde.extras) {
 							m_services->quest_table.progress_quest("npc_dialogue", {label.value().data(), cde.extras->at(0)}, 1, -1, cde.value);
@@ -243,11 +282,12 @@ void Console::handle_inputs(config::ControllerMap& controller) {
 							if (cde.extras->size() > 1) { m_services->quest_table.progress_quest(m_services->quest_registry.get_quest_metadata(cde.value).get_tag(), cde.extras->at(0), cde.extras->at(1)); }
 						}
 					}
-					if (cde.is_piggyback()) { m_services->events.dispatch_event("PiggybackNPC", cde.value); }
-					if (cde.is_open_vendor()) { m_services->events.dispatch_event("OpenVendor", cde.value); }
+					if (cde.is_piggyback()) { m_services->events.npc_piggyback_event.dispatch(*m_services, cde.value); }
+					if (cde.is_open_vendor()) { m_services->events.open_vendor_event.dispatch(*m_services, cde.value); }
 					if (cde.is_item()) {
-						m_services->events.dispatch_event("GivePlayerItem", cde.value, 1);
-						if (cde.extras) { m_services->events.dispatch_event("DestroyInspectable", cde.extras->at(0)); }
+						m_services->events.acquire_item_from_console_event.dispatch(*m_services, cde.value);
+						if (cde.extras) { m_services->data.destroy_inspectable(cde.extras->at(0)); }
+						m_flags.set(ConsoleFlags::close_after_process);
 					}
 					if (cde.is_destructible()) { m_services->data.switch_destructible_state(cde.value); }
 					if (cde.is_exit()) {
@@ -268,7 +308,7 @@ void Console::handle_inputs(config::ControllerMap& controller) {
 			for (auto& code : codes.value()) {
 				if (code.is_response()) {
 					// create a response dialog, feed it inputs, and await its closure before resuming m_writer
-					m_response = ResponseDialog(m_services->text, text_suite, native_key, code.value, m_position + m_response_offset);
+					m_response = ResponseDialog(m_services->text, text_suite, m_services->quest_table, native_key, code.value, m_position + m_response_offset);
 					m_mode = ConsoleMode::responding;
 					m_writer->wait();
 					m_services->soundboard.flags.console.set(audio::Console::next);
@@ -290,16 +330,14 @@ void Console::handle_inputs(config::ControllerMap& controller) {
 					m_writer->set_suite(code.value);
 					if (code.extras) { m_writer->set_index(code.extras.value()[0]); }
 				}
-				if (code.is_start_battle()) { m_services->events.dispatch_event("StartBattle", code.value); }
-				if (code.is_reveal_item() && m_process_codes) { m_services->events.dispatch_event("RevealItem", code.value); }
-				if (code.is_item() && m_process_code_after) { m_services->events.dispatch_event("GivePlayerItem", code.value, 1); }
-				if (code.is_weapon() && m_process_code_after) { m_services->events.dispatch_event("AcquireGun", code.value); }
-				if (code.is_remove_weapon() && m_process_code_after) { m_services->events.dispatch_event("RemovePlayerWeapon", code.value); }
-				if (code.is_destroy_inspectable()) {
-					m_services->data.destroy_inspectable(code.value);
-					m_services->events.dispatch_event("DestroyInspectable", code.value);
-					NANI_LOG_DEBUG(m_logger, "Destroyed Inspectable with ID {}", code.value);
-				}
+				if (code.is_start_battle()) { m_services->events.start_battle_event.dispatch(); }
+				if (code.is_reveal_item() && m_process_codes) { m_services->events.reveal_item_by_id_event.dispatch(code.value); }
+				if (code.is_item() && m_process_code_after) { m_services->events.acquire_item_from_console_event.dispatch(*m_services, code.value); }
+				if (code.is_weapon() && m_process_code_after) { m_services->events.acquire_weapon_from_console_event.dispatch(*m_services, code.value); }
+				if (code.is_remove_weapon() && m_process_code_after) { m_services->events.remove_weapon_by_id_event.dispatch(*m_services, code.value); }
+				if (code.is_open_vendor() && m_process_code_after) { m_services->events.open_vendor_event.dispatch(*m_services, code.value); }
+				if (code.is_emotion() && m_process_code_after && m_npc_portrait && responded) { m_npc_portrait->set_emotion(code.value); }
+				if (code.is_destroy_inspectable()) { m_services->data.destroy_inspectable(code.value); }
 				if (code.is_destructible() && m_process_code_after) {
 					auto inverse = code.extras ? code.extras->at(0) : 0;
 					m_services->data.switch_destructible_state(code.value, static_cast<bool>(inverse));
@@ -310,7 +348,7 @@ void Console::handle_inputs(config::ControllerMap& controller) {
 
 		// go to next message or exit
 		finished = m_writer->request_next();
-		m_process_code_before = finished;
+		m_process_code_before = true;
 		m_process_code_after = true;
 		just_started = finished;
 		can_skip = false;
